+++ back-end/src/lms/serializers/import_schedule.py
@@ -1,5 +1,4 @@
-from django.db import transaction
-from rest_framework import serializers, status, viewsets, generics
+from rest_framework import serializers
 
 from lms.models.lessons import Lesson
 from lms.serializers.lessons import LessonParsedSerializer
@@ -19,12 +18,11 @@ class ImportParsedListSerializer(serializers.Serializer):
     def create(self, validated_data):
         parsed_data = validated_data["lessons"]
 
-        with transaction.atomic():
-            lesson_parsed_objects = []
+        lesson_parsed_objects = []
 
-            for item in parsed_data:
-                item_data = item["parsed"]
-                lesson_parsed_objects.append(Lesson(**item_data))
-            Lesson.objects.bulk_create(lesson_parsed_objects)
+        for item in parsed_data:
+            item_data = item["parsed"]
+            lesson_parsed_objects.append(Lesson(**item_data))
+        Lesson.objects.bulk_create(lesson_parsed_objects)
 
         return validated_data
