+++ back-end/src/lms/views/import_schedule.py
@@ -7,97 +7,16 @@ from auth.models import Permission
 from auth.permissions import BasePermission
 from common.models.subjects import Subject
 from common.parsers import MultiPartWithJSONParser
-import datetime
 from lms.models.common import Milgroup
-
-import openpyxl
-from openpyxl import load_workbook
-import re
-
-from lms.models.lessons import Room, Lesson
+from lms.models.lessons import Lesson, Room
 from lms.models.teachers import Teacher
 from lms.serializers.import_schedule import (
     ImportParsedSerializer,
-    ParseScheduleSerializer,
     ImportParsedListSerializer,
+    ParseScheduleSerializer,
 )
-
+from lms.utils.import_schedule import parse_timetable
 from lms.views.lessons import LessonPermission
-
-month_to_num = {
-    "января": 1,
-    "февраля": 2,
-    "марта": 3,
-    "апреля": 4,
-    "мая": 5,
-    "июня": 6,
-    "июля": 7,
-    "августа": 8,
-    "сентября": 9,
-    "октября": 10,
-    "ноября": 11,
-    "декабря": 12,
-}
-
-MAP_LESSONS_TO_ORDINAL = {"1-2": 1, "3-4": 2, "5-6": 3}
-
-
-def process_date(date):
-    cur_year = datetime.datetime.now().year
-    while "  " in date:
-        date = date.replace("  ", " ")
-    day, month = date.split()
-    date = f"{cur_year}-{month_to_num[month]:02d}-{int(day):02d}"
-    return date
-
-
-def process_field(text):
-    text = text.replace("\n", " ")
-    while "  " in text:
-        text = text.replace("  ", " ")
-
-    # text = text.strip().split(" ")
-    teachers_regexp = re.compile(
-        r"(?P<surname>[А-ЯЁ][а-яё]+) ?"
-        r"(?P<name>[А-ЯЁ])[\.,] ?"
-        r"((?P<patronymic>[А-ЯЁ])[\.,]?)?"
-    )
-    teachers = re.finditer(teachers_regexp, text)
-    teachers = list(
-        map(
-            lambda x: (
-                x.group("surname").lower().strip(),
-                x.group("name").lower().strip(),
-            ),
-            teachers,
-        )
-    )
-
-    if teachers:
-        teachers_start = re.search(teachers_regexp, text).start()
-    else:
-        teachers_start = None
-    classroom = re.search(r"[Аа]уд\.? ?((\d,? ?)*)", text)
-    if not classroom:
-        if "плац" in text.lower():
-            classroom = ["плац"]
-            classroom_start = text.lower().find("плац")
-        else:
-            classroom = []
-            classroom_start = None
-    else:
-        classroom_start = classroom.start()
-        classroom = re.findall(r"\d+", classroom.group())
-        classroom = list(map(str, classroom))
-
-    if classroom_start is not None:
-        subj = text[:classroom_start].strip()
-    else:
-        if teachers_start is not None:
-            subj = text[:teachers_start].strip()
-        else:
-            subj = text.strip()
-    return {"subject": subj, "classrooms": classroom, "teachers": teachers}
 
 
 def guess_subject_name(subject_name):
@@ -109,6 +28,28 @@ def guess_subject_name(subject_name):
         return "Военно-специальная подготовка"
     if "ВПП" in subject_name:
         return "Военно-политическая подготовка"
+    if "ОТ" in subject_name:
+        return "Общая тактика"
+    if "ОП" in subject_name:
+        return "Огневая подготовка"
+    if "ТПЧ" in subject_name:
+        return "Тактика подразделений и частей РВСН"
+    if "ОВУ" in subject_name:
+        return "Общевоинские уставы ВС РФ"
+    if "СтрП" in subject_name:
+        return "Строевая подготовка"
+    if "Тех.П" in subject_name:
+        return "Техническая подготовка"
+    if "УПМВ" in subject_name:
+        return "Управление подразделениями в мирное время"
+    if "РП" in subject_name:
+        return "Разведывательная подготовка"
+    if "ИП" in subject_name:
+        return "Инженерная подготовка"
+    if "РХБЗ" in subject_name:
+        return "Подготовка по РХБЗ"
+    if "СВ" in subject_name:
+        return "Подготовка по связи"
     return ""
 
 
@@ -126,63 +67,6 @@ def guess_subject_type(subject_name):
     if "экзамен" in subject_name.lower():
         return Lesson.Type.EXAM
     return Lesson.Type.OTHER
-
-
-def parse_timetable(timetable_path: str):
-    wb: openpyxl.workbook.Workbook = load_workbook(timetable_path)
-    ws = wb.active
-    x = 0
-    msls = list(ws.merged_cells)
-    for merged_cell in msls:
-        x += 1
-        top_left_cell = ws[merged_cell.coord][0][0]
-        val = top_left_cell.value
-        ws.unmerge_cells(
-            start_row=merged_cell.min_row,
-            start_column=merged_cell.min_col,
-            end_row=merged_cell.max_row,
-            end_column=merged_cell.max_col,
-        )
-        for row in range(merged_cell.min_row, merged_cell.max_row + 1):
-            for col in range(merged_cell.min_col, merged_cell.max_col + 1):
-                cell = ws.cell(row=row, column=col)
-                cell.value = val
-    headers = {}
-    date_header_index = {}
-    row_header = {}
-    for i in range(5, ws.max_column + 1):
-        headers[i] = {}
-        headers[i]["week"] = ws.cell(row=1, column=i).value
-        headers[i]["ordinal"] = ws.cell(row=3, column=i).value
-    now_ind = 2
-    good_rows = []
-    for i in range(4, ws.max_row + 1):
-        if ws.cell(row=i, column=1).value:
-            row_header[i] = {}
-            row_header[i]["week_day"] = ws.cell(row=i, column=1).value
-            row_header[i]["milspec"] = ws.cell(row=i, column=2).value
-            row_header[i]["milgroup"] = ws.cell(row=i, column=3).value
-            date_header_index[i] = now_ind
-            good_rows.append(i)
-        else:
-            now_ind = i
-    final = []
-    for i in range(5, ws.max_column + 1):
-        for elem in good_rows:
-            date = ws.cell(column=i, row=date_header_index[elem]).value
-            ss = ws.cell(row=elem, column=i).value
-            if ss:
-                res = {
-                    **row_header[elem],
-                    **headers[i],
-                    **process_field(ss),
-                    "date": process_date(date),
-                }
-                res["milgroup"] = str(res["milgroup"])
-                if res["ordinal"] in MAP_LESSONS_TO_ORDINAL:
-                    res["ordinal"] = MAP_LESSONS_TO_ORDINAL[res["ordinal"]]
-                    final.append(res)
-    return final
 
 
 class ImportSchedulePermission(BasePermission):
@@ -318,9 +202,6 @@ class ImportParsedView(generics.GenericAPIView):
             serializer = ImportParsedSerializer(data=data)
             if serializer.is_valid(raise_exception=False):
                 valid_data.append(data)
-            else:
-                print("Bad data:")
-                print(data)
         request.data["lessons"] = valid_data
         serializer = self.get_serializer(data=request.data)
         serializer.is_valid(raise_exception=True)
